+++ src/domains/errors/AuthorizationError.ts
@@ -1,0 +1,9 @@
+// src/domain/errors/AuthorizationError.ts
+import { DomainError } from './DomainError';
+
+export class AuthorizationError extends DomainError {
+  constructor(message: string) {
+    super(message);
+    Object.setPrototypeOf(this, AuthorizationError.prototype);
+  }
+}
