+++ src/domains/errors/ConflictError.ts
@@ -1,0 +1,9 @@
+// src/domain/errors/ConflictError.ts
+import { DomainError } from './DomainError';
+
+export class ConflictError extends DomainError {
+  constructor(message: string) {
+    super(message);
+    Object.setPrototypeOf(this, ConflictError.prototype);
+  }
+}
