+++ src/domains/errors/NotFoundError.ts
@@ -1,0 +1,9 @@
+// src/domain/errors/NotFoundError.ts
+import { DomainError } from './DomainError';
+
+export class NotFoundError extends DomainError {
+  constructor(message: string) {
+    super(message);
+    Object.setPrototypeOf(this, NotFoundError.prototype);
+  }
+}
