+++ src/domains/errors/ValidationError.ts
@@ -1,0 +1,9 @@
+// src/domain/errors/ValidationError.ts
+import { DomainError } from './DomainError';
+
+export class ValidationError extends DomainError {
+  constructor(message: string) {
+    super(message);
+    Object.setPrototypeOf(this, ValidationError.prototype);
+  }
+}
