+++ src/domains/errors/index.ts
@@ -1,0 +1,8 @@
+// src/domain/errors/index.ts
+export * from './AuthenticationError';
+export * from './AuthorizationError';
+export * from './ConflictError';
+export * from './DomainError';
+export * from './NotFoundError';
+export * from './ValidationError';
+
