+++ src/server.ts
@@ -82,8 +82,7 @@ if (process.env.NODE_ENV !== 'production') {
   });
 }
 
-const PORT: number = parseInt(process.env.PORT || '3000', 10);
-
+const PORT: number = parseInt(process.env.PORT || '3001', 10);
 app.listen(PORT, () => {
   console.log(`Server running on port ${PORT}`);
   console.log(`Environment: ${process.env.NODE_ENV || 'development'}`);
